+++ core/products/mocks/oracle_engine_mock.go
@@ -35,18 +35,18 @@ func (m *MockOracleEngine) EXPECT() *MockOracleEngineMockRecorder {
 	return m.recorder
 }
 
-// ListensToPubKeys mocks base method.
-func (m *MockOracleEngine) ListensToPubKeys(arg0 oracles.OracleData) bool {
+// ListensToSigners mocks base method.
+func (m *MockOracleEngine) ListensToSigners(arg0 oracles.OracleData) bool {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ListensToPubKeys", arg0)
+	ret := m.ctrl.Call(m, "ListensToSigners", arg0)
 	ret0, _ := ret[0].(bool)
 	return ret0
 }
 
-// ListensToPubKeys indicates an expected call of ListensToPubKeys.
-func (mr *MockOracleEngineMockRecorder) ListensToPubKeys(arg0 interface{}) *gomock.Call {
+// ListensToSigners indicates an expected call of ListensToSigners.
+func (mr *MockOracleEngineMockRecorder) ListensToSigners(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListensToPubKeys", reflect.TypeOf((*MockOracleEngine)(nil).ListensToPubKeys), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListensToSigners", reflect.TypeOf((*MockOracleEngine)(nil).ListensToSigners), arg0)
 }
 
 // Subscribe mocks base method.